requirements.txt
openai
tqdm
faiss-cpu
numpy
pyyaml
pandas
pymupdf
tiktoken
